+++ src/components/dialog/dialog.events.js
@@ -1,0 +1,8 @@
+export default {
+    SHOW: 'dialog-show',
+    AFTER_SHOW: 'dialog-after-show',
+    HIDE: 'dialog-hide',
+    AFTER_HIDE: 'dialog-after-hide',
+    INITIAL_FOCUS: 'dialog-initial-focus',
+    REQUEST_CLOSE: 'dialog-request-close',
+};